Extract compare score logic into its own class
